+++ components/bldc_motor/include/foc_utils.hpp
@@ -6,12 +6,12 @@ namespace espp {
 /**
  * @brief Conversion factor to go from rotations/minute to radians/second.
  */
-static constexpr float RPM_TO_RADS = (2.0f * M_PI) / 60.0f;
+static constexpr float RPM_TO_RADS = (2.0f * (float)(M_PI)) / 60.0f;
 
 /**
  * @brief Conversion factor to go from radians/second to rotations/minute.
  */
-static constexpr float RADS_TO_RPM = 60.0f / (2.0f * M_PI);
+static constexpr float RADS_TO_RPM = 60.0f / (2.0f * (float)(M_PI));
 
 static constexpr float _2_SQRT3 = 1.15470053838f;
 static constexpr float _SQRT3 = 1.73205080757f;
